% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/BiocBook-methods.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{releases}
\alias{chapters}
\title{BiocBook generics}
\usage{
releases(object)

chapters(object)
}
\arguments{
\item{object}{Passed to specific methods}
}
\description{
BiocBook generics
}
