% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/BiocBook-methods.R, R/utils.R
\name{BiocBook-utils}
\alias{BiocBook-utils}
\alias{path,BiocBook-method}
\alias{releases,BiocBook-method}
\alias{chapters,BiocBook-method}
\alias{show,BiocBook-method}
\alias{BiocBook_publish}
\alias{BiocBook_preview}
\alias{BiocBook_versions}
\title{BiocBook utilities}
\usage{
\S4method{path}{BiocBook}(object)

\S4method{releases}{BiocBook}(object)

\S4method{chapters}{BiocBook}(object)

\S4method{show}{BiocBook}(object)

BiocBook_publish(book, message = "🚀 Publishing")

BiocBook_preview(book, browse = FALSE, watch = FALSE)

BiocBook_versions(book)
}
\arguments{
\item{book, object}{A \code{BiocBook} object, created by \code{BiocBook} or \code{BiocBook_init()}.}

\item{message}{Optional. Message used when committing with \code{BiocBook_publish()}.}

\item{browse}{Optional. Passed to \code{quarto_preview()} (default: FALSE).}

\item{watch}{Optional. Passed to \code{quarto_preview()} (default: FALSE).}
}
\description{
BiocBook utilities
}
