% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/utils.R
\name{BiocBook-editing}
\alias{BiocBook-editing}
\alias{edit_book_yml}
\alias{edit_bib}
\alias{edit_requirements_yml}
\title{Editing BiocBook accessory files}
\usage{
edit_book_yml(book, open = TRUE)

edit_bib(book, open = TRUE)

edit_requirements_yml(book, open = TRUE)
}
\arguments{
\item{book}{A \code{BiocBook} object, created by \code{BiocBook} or \code{BiocBook_init()}.}

\item{open}{Optional. Whether to open the file for interactive editing (default: TRUE)}
}
\description{
Editing BiocBook accessory files
}
