% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R, R/BiocBook-methods.R, R/utils.R
\name{BiocBook-utils}
\alias{BiocBook-utils}
\alias{releases}
\alias{chapters}
\alias{path,BiocBook-method}
\alias{releases,BiocBook-method}
\alias{chapters,BiocBook-method}
\alias{show,BiocBook-method}
\alias{BiocBook_publish}
\alias{BiocBook_preview}
\alias{BiocBook_versions}
\title{BiocBook utilities}
\usage{
releases(object)

chapters(object)

\S4method{path}{BiocBook}(object)

\S4method{releases}{BiocBook}(object)

\S4method{chapters}{BiocBook}(object)

\S4method{show}{BiocBook}(object)

BiocBook_publish(book, message = "🚀 Publishing")

BiocBook_preview(book, browse = FALSE, watch = FALSE)

BiocBook_versions(book)
}
\arguments{
\item{book, object}{A \code{BiocBook} object, opened with \code{BiocBook} or created by \code{BiocBook_init()}.}

\item{message}{Optional. Message used when committing with \code{BiocBook_publish()}.}

\item{browse}{Optional. Passed to \code{quarto_preview()} (default: FALSE).}

\item{watch}{Optional. Passed to \code{quarto_preview()} (default: FALSE).}
}
\value{
\itemize{
\item BiocBook_publish: TRUE (invisible) if pushing to Github was successful;
\item BiocBook_preview: Local URL to browse dynamically rendered book;
\item BiocBook_versions: A tibble of the existing versions found on the Github
repository (branch \code{gh-pages}) and of the existing Dockerfiles.
}
}
\description{
BiocBook utilities
}
