% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocBook.R, R/pages.R
\name{BiocBook}
\alias{BiocBook}
\alias{add_preamble}
\title{Handling BiocBook directories}
\usage{
BiocBook(path = ".")

add_preamble(book)
}
\arguments{
\item{path}{Path to the directory of a BiocBook.}

\item{book}{A \code{BiocBook} object, created by \code{BiocBook} or \code{BiocBook_init()}}
}
\description{
The \code{BiocBook()} function creates an S4 object of class \code{BiocBook}.
The \code{BiocBook} S4 class ensures that the input path points towards a
directory matching \code{BiocBook} structure requirements (a \verb{./_quarto.yml} file,
a \verb{./inst/assets/_book.yml}, etc.).

\code{BiocBook} objects can be used to create new chapters with
\code{add_chapter(biocbook, title = "...")}, add a preamble with
\code{add_preamble(biocbook)}, add an appendix with \code{add_appendix()}, ...
}
